+++ packages/yourloops/test/integration/global-setup.js
@@ -25,22 +25,6 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React, { type FC } from 'react'
-import Tooltip from '@mui/material/Tooltip'
-import { usePatientListStyles } from './patient-list.styles'
-
-interface CustomHeaderWithTooltipProps {
-  tooltipText: string
-  headerTitle: string
-}
-
-export const CustomHeaderWithTooltip: FC<CustomHeaderWithTooltipProps> = (props) => {
-  const { tooltipText, headerTitle } = props
-  const { classes } = usePatientListStyles()
-
-  return (
-    <Tooltip title={tooltipText}>
-      <span className={classes.customColumnHeader}>{headerTitle}</span>
-    </Tooltip>
-  )
+module.exports = async () => {
+  process.env.TZ = 'UTC'
 }