+++ packages/yourloops/test/integration/pages/patient-view/daily/daily-view.hcp.spec.tsx
@@ -28,9 +28,8 @@
 import { screen, waitFor } from '@testing-library/react'
 import { mockAuth0Hook } from '../../../mock/auth0.hook.mock'
 import { buildAvailableTeams, mockTeamAPI, myThirdTeamName } from '../../../mock/team.api.mock'
-import { mockDataAPI } from '../../../mock/data.api.mock'
+import { completeDailyViewDataMmol, mockDataAPI } from '../../../mock/data.api.mock'
 import { mockNotificationAPI } from '../../../mock/notification.api.mock'
-import { patient2Id } from '../../../data/patient.api.data'
 import { mockChatAPI } from '../../../mock/chat.api.mock'
 import { mockDirectShareApi } from '../../../mock/direct-share.api.mock'
 import { checkPatientNavBarAsHcp } from '../../../assert/patient-nav-bar.assert'
@@ -40,10 +39,15 @@ import { mockPatientApiForHcp } from '../../../mock/patient.api.mock'
 import { type AppMainLayoutHcpParams, testAppMainLayoutForHcp } from '../../../use-cases/app-main-layout-visualisation'
 import { mockWindowResizer } from '../../../mock/window-resizer.mock'
 import { AppUserRoute } from '../../../../../models/enums/routes.enum'
+import { Unit } from 'medical-domain'
+import { testDailyViewTooltipsAndValuesMmolL } from '../../../use-cases/patient-data-visualisation'
+import { patient2Id } from '../../../data/patient.api.data'
 
 describe('Daily view for HCP', () => {
   const firstName = 'HCP firstName'
   const lastName = 'HCP lastName'
+
+  const dailyRoute = `${AppUserRoute.Patient}/${patient2Id}${AppUserRoute.Daily}`
 
   beforeEach(() => {
     mockWindowResizer()
@@ -74,7 +78,6 @@ describe('Daily view for HCP', () => {
         }
       }
     }
-    const dailyRoute = `${AppUserRoute.Patient}/${patient2Id}${AppUserRoute.Daily}`
     const router = renderPage(dailyRoute)
     await waitFor(() => {
       expect(router.state.location.pathname).toEqual(dailyRoute)
@@ -83,5 +86,21 @@ describe('Daily view for HCP', () => {
     expect(await screen.findByTestId('patient-nav-bar', {}, { timeout: 3000 })).toBeVisible()
     checkPatientNavBarAsHcp()
     await testAppMainLayoutForHcp(appMainLayoutParams)
+  })
+
+  it('should display values in mmol/L according to the HCP settings', async () => {
+    mockUserApi().mockUserDataFetch({
+      firstName,
+      lastName,
+      settings: { units: { bg: Unit.MmolPerLiter } }
+    })
+    mockDataAPI(completeDailyViewDataMmol)
+
+    const router = renderPage(dailyRoute)
+    await waitFor(() => {
+      expect(router.state.location.pathname).toEqual(dailyRoute)
+    })
+
+    await testDailyViewTooltipsAndValuesMmolL()
   })
 })