+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
-## 3.0.5-beta - 2023-07-12
+## 3.0.6-beta - 2023-11-10
+### Changed
+- YLP-2574 Create tool tips independently of the name of the columns
+
+### Engineering use
+- YLP-2476 Use tide-whisperer v2: route /all.
+- YLP-2696 Fix some Sonarcloud bugs and code smells
+- YLP-2704 Remove settings logic from viz
+- YLP-2726 Cleanup tideline/viz packages
+
+## 3.0.5 - 2023-11-02
 ### Added
 - YLP-2292 Display rescue carbs and manual bolus in the trends view for DBLG1 patient
 - YLP-2299 Display DBLG1 events in Daily view
@@ -10,7 +20,6 @@ It is based on Tidepool Blip 1.27.
 ### Changed
 - YLP-2227 Adjustment of the stats on carbs of the dashboard
 - YLP-2484 Remove level column in device change parameter history
-- YLP-2572 Prevent patient from adding the same caregiver to their list twice
 - YLP-2609 Moved patient monitoring alerts configuration to a dedicated view for HCP
 - YLP-2610 Direct link from Monitoring alerts card to "Target & alerts" view
 - YLP-2630 As a user, I would like to be notified with a banner when there are ongoing issues with yourloops
@@ -21,9 +30,11 @@ It is based on Tidepool Blip 1.27.
 - YLP-2670 Remove target & alerts view from "private practice"
 - YLP-2671 Group alarms with the same code occurring in a 30 mn interval
 - YLP-2683 Update alarm device event icon
+- YLP-2690 Switching patient from the tab "Target&Alerts" is not working
 
 ### Fixed
 - YLP-2438 Inconsistency in the daily view between carbs displayed on the graph and Total carbs in the statistics panel
+- YLP-2572 Prevent patient from adding the same caregiver to their list twice
 - YLP-2627 "Last data update" column is not sorting properly
 - YLP-2642 Disabled Google Translate automated translation to avoid React crashes in notification page
 - YLP-2643 Disabled Google Translate automated translation to avoid React crashes in training page
@@ -38,6 +49,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2622 Remove references to uploads in the app
 - YLP-2631 Calls made to /harbour/errors do not always give an error message
 - YLP-2672 Error code given by the front should be findable in kibana
+- YLP-2680 Conversion error in alarm tooltips in mmol/L
 
 ## 3.0.4 - 2023-08-23
 ### Added